+++ packages/jest-worker/src/workers/ChildProcessWorker.ts
@@ -255,6 +255,9 @@ export default class ChildProcessWorker
   }
 
   private _onMessage(response: ParentMessage) {
+    // Ignore messages not intended for us
+    if (!Array.isArray(response)) return;
+
     // TODO: Add appropriate type check
     let error: any;
 
@@ -311,7 +314,8 @@ export default class ChildProcessWorker
         break;
 
       default:
-        throw new TypeError(`Unexpected response from worker: ${response[0]}`);
+        // Ignore messages not intended for us
+        break;
     }
   }